bpf: Fall back to nospec if v1 verification fails
This causes the verifier to fall back to mitigating Spectre v1 using
speculation barriers. A high-level overview of the idea was presented at
LPC'24: https://lpc.events/event/18/contributions/1954/

This requires us to also raise stack/other sanitization errors to
prevent do_check() from falling back to inserting a nospec when on a
speculative path because of the EACCESS.

push_stack() in sanitize_speculative_path() may fail because of resource
exhaustion (including too large bpf progs). In this case we are too
close already, do not try to cover these errors up.

To improve expressiveness further, we also let push_stack() fail for
speculative paths when we are close to the complexity limits. This will
then allow do_check() to insert a nospec instead of verifying a
speculative path and speed up verification. In this case the error is
recoverable.

To distinguish between resource-exhaustion and recoverable errors,
push_stack() must be able to return different error codes and not only a
ptr/NULL. Thus convert it's return value to an error code and store the
state into a output-parameter.

Todos:
- Shorten by using a goto and label for process_bpf_exit() handling
- Remove BUG_ON() calls, left in RFC for assumption-documentation
- Check that ref-leaks due to speculative BPF_EXIT are permitted

Potentioal to split it up further:
- Factor out conversion of push_stack() to error-code returning version
- Factor out introduction of BPF_COMPLEXITY_LIMIT_SPEC_V1_VERIFICATION