bpf: Fall back to nospec if v1 verification fails
This implements the core of VeriFence as presented at LPC'24:
https://lpc.events/event/18/contributions/1954/

This requires us to also raise stack/other sanitization errors.

This prevents do_check() from falling back to inserting a nospec when on
a speculative path because of the EACCESS.

push_stack() in sanitize_speculative_path() will fail because of
resource limits. In this case we are too close already, do not try to
cover these errors up.

To improve expressiveness further, we also let push_stack()
fail for speculative paths when we are close to the complexity limits.
This will then allow do_check() to insert a nospec instead of verifying
a speculative path and speed up verification.

Todos:
- Could likely be shortened further with a goto+label for
  process_bpf_exit() handling