bpf, arm64, powerpc, loongarch: Unset bypass_spec_v1/v4 instead of ignoring BPF_NOSPEC
This changes the semantics of BPF_NOSPEC to always emit a speculation
barrier that works against both Spectre v1 AND v4. If mitigation is not
needed on an architecture, the backend should set
bpf_jit_bypass_spec_v4/v1().

Consequently, sanitize_stack_spill is renamed to nospec_result (meaning
"do not speculate on the resulting state").

This is required for a future commit to allow us to rely on BPF_NOSPEC
from Spectre v4 mitigation for reducing complexity of Spectre v1
speculative path verification.

JITs can set bpf_jit_bypass_spec_v1/v4() if they want the verifier to
skip analysis/patching for the respective vulnerability. For both v1 and
v4, this will reduce the number of barriers the verifier inserts. For
v1, it will (with the future fallback to nospec for v1) also allow more
programs to be accepted in certain edge-cases.

Especially for v1, JIT backends should prefer setting bypass_spec_v1
over simply not lowering BPF_NOSPEC as this will not lead to
false-positive rejections.

*Implementation considerations:*

f5e81d111750 ("bpf: Introduce BPF nospec instruction for mitigating
Spectre v4") noted that mitigation instructions for v1 and v4 might be
different on some archs. For the architectures that implemented
BPF_NOSPEC, this is only the case for powerpc. Even there, some insns
are shared between the barriers for v1 and v4. If this is still found to
impact performance in an unacceptable way, BPF_NOSPEC can be split into
BPF_NOSPEC_V1 and BPF_NOSPEC_V4 later. As an optimization, we can
already skip v1/v4 insns from being emitted for powerpc with this setup
if bypass_spec_v1/v4 is set.

That optimization for powerpc could be further improved by taking the
current value of cpu_mitigations_off() into account. However, this is
not implemented for simplicity as having only v1/v4 should be rare
already and the performance-benefit is currently unknown.

Directly disabling the analysis instead of skipping the lowering of
BPF_NOSPEC might allow programs on non-vulnerable CPUs to be accepted
while the program will be rejected on vulnerable CPUs. With the fallback
to speculation barriers for Spectre v1 implemented in a future commit,
this will only affect programs that do variable stack-accesses or are
very complex.

An alternative would be to always perform the same analysis (not taking
into account whether the current CPU is vulnerable), but this would
needlessly restrict users of CPUs that are not vulnerable. The only
usecase for this would be portability-testing, but this can later be
added easily when needed by allowing users to force bypass_spec_v1/v4 to
false.

*Hardware supported:*

This only supports the architectures already supported since
f5e81d111750 ("bpf: Introduce BPF nospec instruction for mitigating
Spectre v4") and b7540d625094 ("powerpc/bpf: Emit stf barrier
instruction sequences for BPF_NOSPEC"), i.e. x86 (32-bit and 64-bit),
arm64, and powerpc (64-bit). No other architecture has implemented
BPF_NOSPEC since and they may therefore continue to be vulnerable.

a6f6a95f2580 ("LoongArch, bpf: Fix jit to skip speculation barrier
opcode") is the only other past commit related to BPF_NOSPEC and
indicates that the insn is not required there. Consequently, unset
bpf_jit_bypass_spec_v1/v4() for loongarch.

The other archs are left untouched (BPF_NOSPEC is not lowered, but
bypass_spec_v1/v4 is not set) to avoid causing unfounded rejections but
still not implying that they are definately not vulnerable.

*Testing:*

On arm64, 'dsb nsh; isb' was successfully tested with the BPF CI in [1].
'sb' in TODO.

For powerpc, the SEC_FTR checking in bpf_jit_bypass_spec_v4() is based
on the check that was previously located in the BPF_NOSPEC case.

[1] https://github.com/kernel-patches/bpf/pull/8576

Signed-off-by: Luis Gerhorst <luis.gerhorst@fau.de>
Acked-by: Henriette Herzog <henriette.herzog@rub.de>
Cc: Maximilian Ott <ott@cs.fau.de>
Cc: Milan Stephan <milan.stephan@fau.de>

loongarch