bpf: Fall back to nospec for sanitization-failures
For the removed REASON_STACK, this allows us to later fall back to
nospec for certain errors from push_stack() if we are on a speculative
path.

Fall back to nospec directly for the remaining sanitization errs: In
this case we must prevent a following mem-access from using the result
of the alu op speculatively. Therefore, insert a nospec after the alu
insn.

Signed-off-by: Luis Gerhorst <luis.gerhorst@fau.de>