bpf: Fall back to nospec for Spectre v1
This implements the core of the series and causes the verifier to fall
back to mitigating Spectre v1 using speculation barriers. The approach
was presented at LPC'24:
  https://lpc.events/event/18/contributions/1954/ ("Mitigating
  Spectre-PHT using Speculation Barriers in Linux eBPF")
and RAID'24:
  https://arxiv.org/pdf/2405.00078 ("VeriFence: Lightweight and Precise
  Spectre Defenses for Untrusted Linux Kernel Extensions")

In the tests, some are now successful where we previously had a
false-positive (i.e., rejection). Change them to reflect where the
nospec should be inserted (as comment) and modify the error message if
the nospec is able to mitigate a problem that previously shadowed
another problem.

Briefly went through all the occurrences of EPERM, EINVAL, and EACCESS
in the verifier in order to validate that catching them like this makes
sense.

This changes Spectre v1 mitigations to only supporting the architectures
which implement BPF_NOSPEC. Before this commit, Spectre v1 mitigations
prevented exploits by rejecting the programs on all architectures.

It therefore has the downside that Spectre v1 mitigation now requires
BPF_NOSPEC to be emitted by every JIT for archs vulnerable to Spectre
v1. This currently is not the case, and this patch therefore regresses
BPF's security to a limited extent.

The regression is limited to systems vulnerable to Spectre v1, have
unprivileged BPF enabled, and do NOT emit insns for BPF_NOSPEC. The
latter is not the case for x86 64- and 32-bit, arm64, and powerpc 64-bit
and they are therefore not affected by the regression. According to
a6f6a95f2580 ("LoongArch, bpf: Fix jit to skip speculation barrier
opcode"), LoongArch and mips are not vulnerable to Spectre v1 and
therefore also likely not affected by the regression. This contradict [1],
which states that some mips cores may be vulnerable.

Also, if any of those regressed systems is also vulnerable to Spectre
v4, the system was already vulnerable to Spectre v4 attacks based on
unpriv BPF before this patch and the impact is therefore further
limited.

It is unclear which other architectures (besides x86 64- and 32-bit,
arm64, powerpc 64-bit, LoongArch, and mips) supported by the kernel are
vulnerable to Spectre v1 but not to Spectre v4. Also, it is not clear if
barriers are available on these architectures. Implementing BPF_NOSPEC
on these architectures therefore is non-trivial. Searching gcc / the
kernel for speculation barrier implementations for these architectures
yielded no result.

As an alternative to regressing security, one could still reject
programs if the architecture does not emit BPF_NOSPEC (e.g., by removing
the empty BPF_NOSPEC-case from all JITs except for LoongArch and mips
where they appear justified). However, this will cause rejections on
these archs only and this would complicate BPF selftests.

[1] https://wiki.debian.org/DebianSecurity/SpectreMeltdown

Signed-off-by: Luis Gerhorst <luis.gerhorst@fau.de>
Acked-by: Henriette Herzog <henriette.herzog@rub.de>
Cc: Maximilian Ott <ott@cs.fau.de>
Cc: Milan Stephan <milan.stephan@fau.de>